requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
google-genai>=1.0.0
networkx>=3.3
scipy>=1.13.0
numpy>=1.26.0
cvxpy>=1.5.0
pydantic>=2.7.0
python-dotenv>=1.0.0
